requirements.txt
Django==2.2.5
django-dynamic-raw-id==2.6
django-admin-rangefilter==0.5.0

ipython==7.8.0

psycopg2-binary==2.8.3
